omit susp in accomp; pattern fallbacks; etc.

+++ dumb_composer/__main__.py
@@ -9,6 +9,8 @@ from dumb_composer.dumb_composer import PrefabComposer
 
 
 def configure_logging(log_file_path, log_level, append_to_log):
+    if log_file_path is None:
+        return
     if not append_to_log:
         if os.path.exists(log_file_path):
             os.remove(log_file_path)
